Title: a/b 출력, Time: -1ms, Memory: -1MB, Status: Compile Error - CodeTree

+++ 250117/a/b 출력/a-divide-b.java
@@ -10,7 +10,7 @@ public class Main {
         a = sc.nextInt();
         b = sc.nextInt();
 
-        answer = (int)(a/b*10000000000)/10000000000.0;
+        answer = (a/b*10000000000)/10000000000.0;
         System.out.print(answer);
     }
 }
